pyproject.toml
[build-system]
requires = [
    "setuptools>=42",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "tycho-api"
authors = [
    {name = "Renaissance Computing Institute", email = "scox@renci.org"}
]
maintainers = [
    {name = "Murali Karthik Kothapalli", email = "muralikarthik.k@renci.org"}
]
requires-python = ">=3.7"
description = "Tycho is an API, compiler, and executor for cloud native distributed systems."
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
dependencies = [
    "deepmerge==1.0.1",
    "flasgger==0.9.5",
    "Flask==2.0.3",
    "flask_cors==3.0.10",
    "flask_restful==0.3.9",
    "GitPython==3.1.11",
    "Jinja2==3.0.3",
    "jsonschema==3.2.0",
    "kubernetes==25.3.0",
    "netifaces==0.11.0",
    "PyYAML==5.4.1",
    "requests==2.31.0",
    "docker-compose==1.29.2",
    "requests_cache==0.9.2",
]
dynamic = ["version"]

[project.urls]
documentation = "https://github.com/helxplatform/tycho"
repository = "https://github.com/helxplatform/tycho"

[project.scripts]
tycho = "tycho.api:main"

[tool.setuptools]
packages = ["tycho", "tycho.template", "tycho.conf"] 
include-package-data = true
zip-safe = false

[tool.setuptools.package-data]
tycho = ["*yaml"]
"tycho.conf" = ["*.yaml"]
"tycho.template" = ["*.yaml"]

[tool.setuptools.dynamic]
version = {attr = "tycho.VERSION"}

